Call preg_match with '0' as flags-parameter
hhvm croaks if preg_match is called with `null` as flags-parameter, because it expects a numerical value. Replacing `null` with `0` in the calls to preg_match that violate this expectation allows smarty to be used under hhvm; we have a mixed setup, running under php-fpm as well as under hhvm on the commandline for tools needing the extra performance and generally saner behaviour of hhvm.

I would be pleased if you would consider incorporating this small fix, or educate me if my fix is violating some standard or good-practice I am unaware of.

Many thanks for the great work on smarty -- it is much appreciated.