Fix foreachelse with foreach on objects
Try to address issue #506. After this patch the foreach logic always
count the total number of elements when the underlying structure can
easily be queried. E.g.: arrays and Countables can be queried with count()
for the total number of elements. For all iterators we do not
know the total element count in advance and calculating it is a wasteful
operation. And it can be dangerous, too: forward-iterators cannot be
rewinded and the foreach-loop will break. So with this patch the @total
property will be simply set to -1 for iterators.
This has implications to other properties as well, when foreach is
used with iterarors: The @show property is
always true (even when there are no elements), and @last is always false.

This patch changes a runtime function call. So all templates that uses
foreach must be recompiled.